Fix tests issues with Grails 2.0.0.RC1
