+++ application.properties
@@ -1,5 +1,5 @@
 #Grails Metadata file
-#Tue Oct 25 19:32:03 CEST 2011
+#Thu Nov 10 11:53:54 CET 2011
 app.grails.version=2.0.0.RC1
 app.name=jaxrs
 app.servlet.version=2.5
